requirements.txt
charset-normalizer==3.4.3
cssselect2==0.8.0
hmmlearn==0.3.2
lxml==6.0.1
numpy==2.3.2
pandas==2.3.2
pillow==11.3.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
reportlab==4.4.3
six==1.17.0
svglib==1.5.1
tinycss2==1.4.0
tzdata==2025.2
webencodings==0.5.1
